+++ dune/gfe/densities/duneelasticitydensity.hh
@@ -16,18 +16,22 @@ namespace Dune::GFE
 
   /** \brief Adapter for densities from the dune-elasticity module
    *
+   * \tparam ElementOrIntersection The domain of the density.
+   *   Can be either a grid element (i.e., a codimension-0 Entity)
+   *   or an Intersection.
    * \tparam index The TargetSpace is typically a product with one factor being
    *   RealTuple. The 'index' argument selects which factor to assign the
    *   dune-elasticity density to.
    */
-  template<class Position, class TargetSpace, std::size_t index=0>
+  template<class ElementOrIntersection, class TargetSpace, std::size_t index=0>
   class DuneElasticityDensity final
-    : public LocalDensity<Position,TargetSpace>
+    : public LocalDensity<ElementOrIntersection,TargetSpace>
   {
-    using ctype = typename Position::value_type;
+    using LocalCoordinate = typename ElementOrIntersection::Geometry::LocalCoordinate;
+    using ctype = typename ElementOrIntersection::Geometry::ctype;
     using field_type = typename TargetSpace::field_type;
 
-    constexpr static auto dim = Position::size();
+    constexpr static auto dim = LocalCoordinate::size();
     constexpr static auto embeddedBlocksize = TargetSpace::EmbeddedTangentVector::dimension;
 
     using ATargetSpace = typename TargetSpace::template rebind<adouble>::other;
@@ -52,7 +56,7 @@ namespace Dune::GFE
      * \param value The deformation at the current position
      * \param derivative The derivative at the current position
      */
-    virtual field_type operator() (const Position& x,
+    virtual field_type operator() (const LocalCoordinate& x,
                                    const typename TargetSpace::CoordinateType& value,
                                    const FieldMatrix<field_type,embeddedBlocksize,dim>& derivative) const override
     {
@@ -69,13 +73,13 @@ namespace Dune::GFE
     }
 
     // Construct a copy of this density but using 'adouble' as the number type
-    virtual std::unique_ptr<LocalDensity<Position,ATargetSpace> > makeActiveDensity() const
+    virtual std::unique_ptr<LocalDensity<ElementOrIntersection,ATargetSpace> > makeActiveDensity() const
     {
       // The active dune-elasticity density
       auto activeDensity = elasticityDensity_->makeActiveDensity();
 
       // Wrap it as a dune-gfe density
-      return std::make_unique<DuneElasticityDensity<Position,ATargetSpace,index> >(std::move(activeDensity));
+      return std::make_unique<DuneElasticityDensity<ElementOrIntersection,ATargetSpace,index> >(std::move(activeDensity));
     }
 
     /** \brief The density does not depend on the value */