+++ dune/gfe/densities/planarcosseratshelldensity.hh
@@ -21,14 +21,15 @@
 
 namespace Dune::GFE
 {
-  template<class Position, class field_type>
+  template<class ElementOrIntersection, class field_type>
   class PlanarCosseratShellDensity final
-    : public GFE::LocalDensity<Position, GFE::ProductManifold<RealTuple<field_type,3>,Rotation<field_type,3> > >
+    : public GFE::LocalDensity<ElementOrIntersection, GFE::ProductManifold<RealTuple<field_type,3>,Rotation<field_type,3> > >
   {
   private:
+    using LocalCoordinate = typename ElementOrIntersection::Geometry::LocalCoordinate;
     // PlanarCosseratShellDensity only works for 2d grids
-    static_assert(Position::size()==2);
-    static constexpr int gridDim = Position::size();
+    static_assert(LocalCoordinate::size()==2);
+    static constexpr int gridDim = LocalCoordinate::size();
 
     // The target space with 'adouble' as the number type
     using ATargetSpace = GFE::ProductManifold<RealTuple<adouble,3>,Rotation<adouble,3> >;
@@ -210,7 +211,7 @@ namespace Dune::GFE
 
     /** \brief Evaluate the density
      */
-    virtual field_type operator() (const Position& x,
+    virtual field_type operator() (const LocalCoordinate& x,
                                    const typename GFE::ProductManifold<RealTuple<field_type,3>,Rotation<field_type,3> >::CoordinateType& value,
                                    const FieldMatrix<field_type,7,gridDim>& derivative) const override
     {
@@ -253,9 +254,9 @@ namespace Dune::GFE
     }
 
     // Construct a copy of this density but using 'adouble' as the number type
-    virtual std::unique_ptr<LocalDensity<Position,ATargetSpace> > makeActiveDensity() const override
+    virtual std::unique_ptr<LocalDensity<ElementOrIntersection,ATargetSpace> > makeActiveDensity() const override
     {
-      return std::make_unique<PlanarCosseratShellDensity<Position,adouble> >(thickness_, mu_, lambda_, mu_c_, L_c_, q_, kappa_);
+      return std::make_unique<PlanarCosseratShellDensity<ElementOrIntersection,adouble> >(thickness_, mu_, lambda_, mu_c_, L_c_, q_, kappa_);
     }
 
     /** \brief The density depends on the microrotation value, but not on the deformation